fix(review): centralize GitHub review test builders
